pyproject.toml
[project]
name = "ma-mpd"
version = "0.1.0"
description = "Multi-Agent Python Package Detection - Malicious Package Classification System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core framework
    "agents>=1.4.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    
    # AI/ML and agent frameworks
    "langchain>=0.0.350",
    "langsmith>=0.0.69",
    "openai>=1.3.7",
    "mlflow>=2.8.1",
    
    # Database
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    
    # Data processing
    "pandas>=2.1.4",
    "numpy>=1.24.4",
    "pydantic>=2.5.1",
    
    # Archive handling and security
    "py7zr>=0.20.8",
    
    # Configuration and environment
    "python-dotenv>=1.0.0",
    "configparser>=6.0.0",
    
    # Logging and monitoring
    "structlog>=23.2.0",
    "python-json-logger>=2.0.7",
    "psutil>=5.9.6",
    
    # File handling
    "pathlib2>=2.3.7",
    "aiofiles>=23.2.1",
    
    # Testing
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    
    # Utilities
    "typing-extensions>=4.8.0",
]
